Better way to configure unique terms within an expression (#775)
* Push the configuration for the UniqueExpressionTermsVisitor down to the logic level vs. the DefaultQueryPlanner level.

* Don't drop config options when tuning lookupUUID.

* Remove UniqueExpressionVisitor hook form DefaultQueryPlanner, rely solely on the shardQueryConfig.

* Only set config option once.
Conflicts:
	warehouse/query-core/src/main/java/datawave/query/config/ShardQueryConfiguration.java
	warehouse/query-core/src/test/java/datawave/query/config/ShardQueryConfigurationTest.java
