+++ warehouse/query-core/src/main/java/datawave/query/planner/DefaultQueryPlanner.java
@@ -149,11 +149,6 @@ public class DefaultQueryPlanner extends QueryPlanner {
     protected boolean limitScanners = false;
     
     /**
-     * Allows developers to enable/disable the enforcing of unique nodes with OR and AND nodes.
-     */
-    private boolean enforceUniqueTermsWithinExpressions = false;
-    
-    /**
      * Allows developers to disable bounded lookup of ranges and regexes. This will be optimized in future releases.
      */
     protected boolean disableBoundedLookup = false;
@@ -739,7 +734,8 @@ public class DefaultQueryPlanner extends QueryPlanner {
         
         stopwatch.stop();
         
-        if (enforceUniqueTermsWithinExpressions) {
+        // Enforce unique terms within an AND or OR expression.
+        if (config.getEnforceUniqueTermsWithinExpressions()) {
             stopwatch = timers.newStartedStopwatch("DefaultQueryPlanner - Enforce unique terms within AND and OR expressions");
             queryTree = UniqueExpressionTermsVisitor.enforce(queryTree);
             if (log.isDebugEnabled()) {
@@ -2438,13 +2434,5 @@ public class DefaultQueryPlanner extends QueryPlanner {
         if (null != builderThread) {
             builderThread.shutdown();
         }
-    }
-    
-    public boolean isEnforceUniqueTermsWithinExpressions() {
-        return enforceUniqueTermsWithinExpressions;
-    }
-    
-    public void setEnforceUniqueTermsWithinExpressions(boolean enforceUniqueTermsWithinExpressions) {
-        this.enforceUniqueTermsWithinExpressions = enforceUniqueTermsWithinExpressions;
     }
 }
