+++ warehouse/query-core/src/main/java/datawave/query/tables/ShardQueryLogic.java
@@ -1931,6 +1931,14 @@ public class ShardQueryLogic extends BaseQueryLogic<Entry<Key,Value>> {
         this.config.setCollapseUidsThreshold(collapseUidsThreshold);
     }
     
+    public boolean getEnforceUniqueTermsWithinExpressions() {
+        return this.config.getEnforceUniqueTermsWithinExpressions();
+    }
+    
+    public void setEnforceUniqueTermsWithinExpressions(boolean enforceUniqueTermsWithinExpressions) {
+        this.getConfig().setEnforceUniqueTermsWithinExpressions(enforceUniqueTermsWithinExpressions);
+    }
+    
     public long getMaxIndexScanTimeMillis() {
         return getConfig().getMaxIndexScanTimeMillis();
     }
