+++ warehouse/query-core/src/test/java/datawave/query/config/ShardQueryConfigurationTest.java
@@ -430,7 +430,7 @@ public class ShardQueryConfigurationTest {
      */
     @Test
     public void testCheckForNewAdditions() throws IOException {
-        int expectedObjectCount = 165;
+        int expectedObjectCount = 166;
         ShardQueryConfiguration config = ShardQueryConfiguration.create();
         ObjectMapper mapper = new ObjectMapper();
         JsonNode root = mapper.readTree(mapper.writeValueAsString(config));
